+++ tests/canary.rs
@@ -1,4 +1,4 @@
-pub fn add(left: usize, right: usize) -> usize {
+pub fn add(left: u32, right: u32) -> u32 {
     left + right
 }
